app/requirements.txt
Django==5.1
django-cors-headers==4.3.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangorestframework-camel-case==1.4.2
django-environ==0.11.2
django-imagekit==5.0.0
djoser==2.3.1
drf-spectacular==0.28.0
django-unfold==0.51.0
Pillow==10.4.0
psycopg2-binary~=2.9.10
gunicorn==23.0.0
